Move MountDevice type declaration to @wp-playground/storage
